Remove dtype argument from VarianceScaling initializer

+++ VishwamAI/scripts/model_architecture.py
@@ -50,7 +50,7 @@ class VishwamAIModel(hk.Module):
         ) for _ in range(self.num_experts)]
 
         # Define gating mechanism
-        self.gating_network = hk.Linear(self.num_experts, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", "uniform", dtype=jnp.float32))
+        self.gating_network = hk.Linear(self.num_experts, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", "uniform"))
 
     def __call__(self, inputs):
         if isinstance(inputs, str):
